Fixes/improvements.
Fixed a bug in the "DHCP Lease Time" input field where entries beginning with ZERO such as "0120" or "012d" were considered valid.
Tightening the code that checks the format of the special syntax notation that include valid suffixes (s, m, h, d, w).
Added code to set the focus on the input field that fails its validation when the "Apply" button is clicked.